Move the cart selectors to beside the cart reducer.
