+++ web/app/containers/mini-cart/partials/mini-cart-product-list.jsx
@@ -9,7 +9,7 @@ import Image from 'progressive-web-sdk/dist/components/image'
 import List from 'progressive-web-sdk/dist/components/list'
 import ProductItem from '../../../components/product-item'
 
-import * as selectors from '../selectors'
+import * as selectors from '../../../store/cart/selectors'
 
 // Parses strings in the format: $Dollars.Cents
 // Dollar amounts only, cents must be specified.
